+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 The format is modified from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### 2023.04.01
+
+##### [v0.1.5] - Recognize Milliliter
+**Added**
+- 新增生命設定與提示
+  
+
+### 2023.03.30
+
+##### [v0.0.0] - Educational Games
+**Changed**
+- 更新網站顯示樣式
+  
+##### [v0.1.1] - Recognize Angles
+**Added**
+- 新增提示
+  
 ### 2023.03.29
 
 ##### [v0.1.0] - Recognize Angles
@@ -19,7 +36,7 @@ The format is modified from [Keep a Changelog](https://keepachangelog.com/en/1.0
 **Changed**
 - 更新提示畫面，顯示水杯的容量
 
-##### [v0.1.0] - Game Tempalte
+##### [v0.1.0] - Game Template
 **Changed**
 - 更新功能使用責任鍊模式，和建立遊戲類別
   
@@ -34,11 +51,9 @@ The format is modified from [Keep a Changelog](https://keepachangelog.com/en/1.0
 - 支援觸控遊玩方式。
 
 **Fixed**
-- Connection Milliliter
-  - 取消默認的滾動頁面
+- 取消默認的滾動頁面
 
 ##### [v0.1.4] - Recgnize Milliliter
 
 **Fixed**
-- Recgnize Milliliter
-  - 取消默認的滾動頁面
+- 取消默認的滾動頁面